review: split taste-profile.service.ts into pipelines + queries (ROK-948)
Review caught that taste-profile.service.ts was 704 lines — 2.3x the
300-line ESLint max-lines cap (warn, heading to error per CLAUDE.md).
Only other violator in api/src/**/*.service.ts was settings at 400;
704 was a clear outlier. Design-small-from-the-start goal missed — fixing
before the PR lands rather than queuing as tech debt.

Split into focused files, each under 300 lines:
- pipelines/aggregate-vectors.ts (254) — daily signal → vector pipeline
  plus its private helpers (loadUserSignals, loadGameMetadata,
  computeUserSignalHash, coPlayPartnerCount)
- pipelines/weekly-intensity.ts (171) — weekly rollup pipeline +
  currentWeekStart, maxUniqueGamesThisWeek, buildWeeklySnapshot
- pipelines/build-co-play-graph.ts — voice overlap + signup pair builder
- queries/taste-profile-queries.ts (167) — getTasteProfile,
  findSimilarPlayers, topCoPlayPartners, zeroedDimensions
- taste-profile.service.ts (now ~80 lines) — thin NestJS wrapper: cron
  decorators + one-line delegators to the pure functions

Everything public keeps the same shape so tests are unchanged.
All 58 tests still green; local CI 7/7 pass.
